Remove the interop files from the build
