+++ src/libraries/Common/src/Interop/Unix/System.Net.Security.Native/Interop.NetSecurityNative.cs
@@ -169,51 +169,5 @@ internal static partial class Interop
 
             return Unwrap(out minorStatus, contextHandle, inputBytes, offset, count, ref outBuffer);
         }
-
-        // https://www.gnu.org/software/gss/reference/gss.pdf Page 65
-        internal const int GSS_C_ROUTINE_ERROR_OFFSET = 16;
-
-        // https://www.gnu.org/software/gss/reference/gss.pdf Page 9
-        internal enum Status : uint
-        {
-            GSS_S_COMPLETE = 0,
-            GSS_S_CONTINUE_NEEDED = 1,
-            GSS_S_BAD_MECH = 1 << GSS_C_ROUTINE_ERROR_OFFSET,
-            GSS_S_BAD_NAME = 2 << GSS_C_ROUTINE_ERROR_OFFSET,
-            GSS_S_BAD_NAMETYPE = 3 << GSS_C_ROUTINE_ERROR_OFFSET,
-            GSS_S_BAD_BINDINGS = 4 << GSS_C_ROUTINE_ERROR_OFFSET,
-            GSS_S_BAD_STATUS = 5 << GSS_C_ROUTINE_ERROR_OFFSET,
-            GSS_S_BAD_SIG = 6 << GSS_C_ROUTINE_ERROR_OFFSET,
-            GSS_S_NO_CRED = 7 << GSS_C_ROUTINE_ERROR_OFFSET,
-            GSS_S_NO_CONTEXT = 8 << GSS_C_ROUTINE_ERROR_OFFSET,
-            GSS_S_DEFECTIVE_TOKEN = 9 << GSS_C_ROUTINE_ERROR_OFFSET,
-            GSS_S_DEFECTIVE_CREDENTIAL = 10 << GSS_C_ROUTINE_ERROR_OFFSET,
-            GSS_S_CREDENTIALS_EXPIRED = 11 << GSS_C_ROUTINE_ERROR_OFFSET,
-            GSS_S_CONTEXT_EXPIRED = 12 << GSS_C_ROUTINE_ERROR_OFFSET,
-            GSS_S_FAILURE = 13 << GSS_C_ROUTINE_ERROR_OFFSET,
-            GSS_S_BAD_QOP = 14 << GSS_C_ROUTINE_ERROR_OFFSET,
-            GSS_S_UNAUTHORIZED = 15 << GSS_C_ROUTINE_ERROR_OFFSET,
-            GSS_S_UNAVAILABLE = 16 << GSS_C_ROUTINE_ERROR_OFFSET,
-            GSS_S_DUPLICATE_ELEMENT = 17 << GSS_C_ROUTINE_ERROR_OFFSET,
-            GSS_S_NAME_NOT_MN = 18 << GSS_C_ROUTINE_ERROR_OFFSET,
-        }
-
-        [Flags]
-        internal enum GssFlags : uint
-        {
-            GSS_C_DELEG_FLAG = 0x1,
-            GSS_C_MUTUAL_FLAG = 0x2,
-            GSS_C_REPLAY_FLAG = 0x4,
-            GSS_C_SEQUENCE_FLAG = 0x8,
-            GSS_C_CONF_FLAG = 0x10,
-            GSS_C_INTEG_FLAG = 0x20,
-            GSS_C_ANON_FLAG = 0x40,
-            GSS_C_PROT_READY_FLAG = 0x80,
-            GSS_C_TRANS_FLAG = 0x100,
-            GSS_C_DCE_STYLE = 0x1000,
-            GSS_C_IDENTIFY_FLAG = 0x2000,
-            GSS_C_EXTENDED_ERROR_FLAG = 0x4000,
-            GSS_C_DELEG_POLICY_FLAG = 0x8000
-        }
     }
 }
