+++ src/libraries/Common/src/System/Net/ContextFlagsAdapterPal.PlatformNotSupported.cs
@@ -1,0 +1,20 @@
+// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+
+using System;
+
+namespace System.Net
+{
+    internal static class ContextFlagsAdapterPal
+    {
+        internal static ContextFlagsPal GetContextFlagsPalFromInterop(Interop.NetSecurityNative.GssFlags gssFlags, bool isServer)
+        {
+            throw new PlatformNotSupportedException();
+        }
+
+        internal static Interop.NetSecurityNative.GssFlags GetInteropFromContextFlagsPal(ContextFlagsPal flags, bool isServer)
+        {
+            throw new PlatformNotSupportedException();
+        }
+    }
+}
